Finished,
When at the properties nav, the user is given a list of properties with descriptions. Upon choosing the item to select, id will grab the id associated and take the user to a page with photos. Within this page is also the functionality to set the order of images the user would want.